pyproject.toml
[tool.poetry]
name = "nonebot-plugin-multimodal-gemini"
version = "0.0.3"
description = "Nonebot2 的谷歌 Gemini 多模态助手，一个命令即可玩转 Gemini 的多模态！"
authors = ["zhiyu1998 <renzhiyu0416@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot-plugin-multimodal-gemini"}]
repository = "https://github.com/zhiyu1998/nonebot-plugin-multimodal-gemini"
keywords = ["nonebot", "nonebot2", "gemini"]

[tool.poetry.dependencies]
python = ">=3.9"
aiohttp = "^3.7"
httpx = ">=0.23"
nonebot2 = ">=2.3.0"
aiofiles = ">=0.8.0"
google-generativeai = ">=0.8.3"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
nonebot_plugin_apscheduler = "^0.5.0"
nonebot_plugin_localstore = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"